Updates to collaboration screen
