--- src/apps/project-collaboration/InviteListOfUsers/index.ts
@@ -1,1 +1,0 @@
-export * from './InviteListOfUsers';
