+++ src/apps/project-collaboration/InviteListOfUsersDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './InviteListOfUsersDialog';
