--- src/apps/project-collaboration/InviteUser/index.ts
@@ -1,1 +1,0 @@
-export * from './InviteUser';
