+++ src/apps/project-collaboration/InviteUserDialog/InviteUserDialog.css
@@ -1,0 +1,12 @@
+.invite-user-item {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding: 10px;
+  gap: 5px;
+  font-size: 13px;
+  font-weight: 400;
+  line-height: 18.2px;
+  color: black;
+  cursor: pointer;
+}
