+++ src/apps/project-collaboration/InviteUserDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './InviteUserDialog';
