+++ src/components/DropdownButton/DropdownButton.css
@@ -1,0 +1,3 @@
+.dropdown-button-trigger {
+  gap: 5px;
+}
