+++ src/i18n/de/project-collaboration.json
@@ -60,5 +60,6 @@
   "Invitation List": "Benutzerliste",
   "Send Invitations": "Einladungen Senden",
   "Invitations Sent": "Einladungen Gesendet",
-  "Invitations were sent to your list of users.": "Die Einladungen wurden an die Benutzer versandt."
+  "Invitations were sent to your list of users.": "Die Einladungen wurden an die Benutzer versandt.",
+  "Add single user": "Add single user"
 }
