+++ src/i18n/en/project-collaboration.json
@@ -60,5 +60,6 @@
   "Invitation List": "Invitation List",
   "Send Invitations": "Send Invitations",
   "Invitations Sent": "Invitations Sent",
-  "Invitations were sent to your list of users.": "Invitations were sent to your list of users."
+  "Invitations were sent to your list of users.": "Invitations were sent to your list of users.",
+  "Add single user": "Add single user"
 }
